requirements.txt
pandas~=2.3.1
chromadb~=1.3.4
groq~=0.34.0
python-dotenv~=1.2.1
streamlit~=1.48.0
lxml~=6.0.2
beautifulsoup4~=4.13.5
python-dotenv
sentence_transformers
pandasql
semantic-router[local]